SimpleMessage: add a new endpoint for sending messages.
This will probably be replaced with a more robust mechanism that supports
a better encryption mechanisms. This is just here for the meanwhile to
support pressing needs for a way to message users.

Once this is replaced, we will probably also replace the invitation
mechanism to use the new messaging mechanism.