Gracefully handle uploading the same revision
This is needed so that immediately re-played requests don't fail.

Consider for example the following scenario: a client makes a batch
request that registers correctly on the server, but fails to return
(e.g. a networking error after the request has been processed). The
client would think that the request failed, but the server will already
have the up to date information. This commit just returns a successful
status if this request is sent again (by the client retrying the
request) instead of returning a conflict.

This however doesn't handle the case of a request failing, a
modification being made by another client, and then the request being
retried. This case will stay fail.