Update dependecies generated by pip-compile. (#126)
This one is to fix my own PR #125, the requirement files contained the following sentence:

"To update, run: pip-compile --output-file=requirements.txt requirements.in/base.txt"

But that was misleading, after reading pip-tools documentation I found that

"If pip-compile finds an existing requirements.txt file that fulfils the dependencies then no changes will be made, even if updates are available."

That was my mistake, generated the files again and made new builds using python 3.10 that worked as expected. Once again sorry for the mistake!